fix: Remove per-job HeaderFilter.
It is passed when the worker is initialized; there is no
need to pass it again for every job.
